requirements.txt
discord.py
requests
python-dotenv
python-telegram-bot